+++ previous-versions/accounting/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.31.0] previous-versions/accounting
 ### Releases
 - [PyPI v0.31.0] https://pypi.org/project/codat-accounting/0.31.0 - previous-versions/accounting
+
+## 2023-11-06 01:25:47
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.32.0] previous-versions/accounting
+### Releases
+- [PyPI v0.32.0] https://pypi.org/project/codat-accounting/0.32.0 - previous-versions/accounting